+++ src/components/layouts/mobile-view.tsx
@@ -4,21 +4,17 @@ import { useClip } from "@/contexts"
 import { useScrollY } from "@/hooks"
 import {
   AspectRatio,
-  Box,
-  Skeleton,
   TabsProps,
-  useBoolean,
   useWindowEvent,
   VStack,
 } from "@yamada-ui/react"
-import { useEffect, useMemo, useRef, useState } from "react"
+import { useMemo, useRef, useState } from "react"
 
 export function MobileView() {
   const tabsRef = useRef<HTMLDivElement>()
   const { currentClip } = useClip()
   const y = useScrollY()
   const [width, setWidth] = useState(window?.innerWidth ?? 0)
-  const [isMounted, { off, on }] = useBoolean()
 
   const { height = 0 } = tabsRef.current?.getBoundingClientRect() ?? {}
   const isScroll = y > height
@@ -41,17 +37,18 @@ export function MobileView() {
 
   const src = `${currentClip?.embed_url}&parent=localhost&parent=www.twitchclipsranking.com&parent=twitchclipsranking.com`
 
-  useEffect(off, [off, src])
-
   return (
     <VStack gap={1}>
-      <Box position="sticky" top={0} zIndex="jeice">
-        <Skeleton isFitContent isLoaded={isMounted} w="full">
-          <AspectRatio apply="layoutStyles.borderCard" ratio={16 / 9} w="full">
-            <iframe allowFullScreen loading="lazy" onLoad={on} src={src} />
-          </AspectRatio>
-        </Skeleton>
-      </Box>
+      <AspectRatio
+        apply="layoutStyles.player"
+        position="sticky"
+        ratio={16 / 9}
+        top={0}
+        w="full"
+        zIndex="jeice"
+      >
+        <iframe allowFullScreen loading="lazy" src={src} />
+      </AspectRatio>
 
       <ClipInfo />