+++ src/components/layouts/pc-view.tsx
@@ -25,7 +25,7 @@ export function PCView() {
     >
       <GridItem colSpan={9} w="full">
         <VStack gap={2} marginTop="md">
-          <AspectRatio apply="layoutStyles.borderCard" ratio={16 / 9} w="full">
+          <AspectRatio apply="layoutStyles.player" ratio={16 / 9} w="full">
             <iframe allowFullScreen loading="lazy" src={src} />
           </AspectRatio>